refactor(engine): move the unlinked-issue candidate pre-filter into the shared engine
The deterministic unlinked-issue candidate pre-filter (findUnlinkedIssueCandidates) is
confirmed-pure -- no IO, no AI, no other-src imports -- so it moves cleanly into
@loopover/engine (#4883), letting the published miner/mcp CLIs run the identical recall
pass the maintainer gate uses instead of maintaining a driftable second copy.

- new packages/loopover-engine/src/signals/unlinked-issue-candidates.ts holds the canonical
  implementation.
- src/signals/unlinked-issue-candidates.ts becomes a thin re-export shim, imported via the
  relative source path (matching src/signals/slop.ts) so check-engine-parity recognizes it as
  a shim rather than a hand-duplicated twin, and so it never depends on the engine's built dist/.
- re-exported from the engine barrel for external @loopover/engine consumers.

No behavior change: the existing test/unit/unlinked-issue-candidates.test.ts imports the shim
path unchanged and passes (100% line/branch coverage of the moved module).

Closes #4883
